+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.3.2] / 2025-03-05
+### Features
+- Support localized `InstallLocation` path.
+### Updates
+- Add `FinalPathNameExtension` to fix localized `InstallLocation`. (Fix: #13)
+- Update `GetProcesses` to use `FinalPathNameExtension` to fix localized `InstallLocation`.
+- Update typo related to `Journal`.
+- Update `ApplicationPluginsUtils` to fix `bundle.zip` download.
+### Tests
+- Add `FinalPathName_Tests` with `english`, `portuguese` and `german` tests.
+- Add `BundleCreatorUtils` to test `ApplicationPluginsUtils` with fake `bundle.zip`.
+
 ## [1.3.1] / 2024-12-16
 ### Features
 - `ApplicationPluginsUtils` with `Mutex` to prevent multiple instances. (Fix: #11)
@@ -82,6 +94,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `ricaun.Revit.Installation.Tests`
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.3.2]: ../../compare/1.3.1...1.3.2
 [1.3.1]: ../../compare/1.3.0...1.3.1
 [1.3.0]: ../../compare/1.2.0...1.3.0
 [1.2.0]: ../../compare/1.1.2...1.2.0